pyproject.toml
[project]
name = "github-reader-mcp"
version = "0.1.0"
description = "GitHub Reader MCP Server - Observe public repos via GraphQL"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.1.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
]
